requirements.txt
bleach==1.5.0
html5lib==0.9999999
Markdown==2.6.9
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
protobuf==3.4.0
scikit-learn==0.19.0
scipy==0.19.1
six==1.11.0
sklearn==0.0
tensorflow==1.3.1
tensorflow-tensorboard==0.1.8
Werkzeug==0.12.2
